backend/requirements.txt
# Flask and web framework dependencies
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Limiter==3.5.0

# Background job processing
celery==5.3.6
redis==5.0.1

# File handling and validation
Werkzeug==3.0.3
python-magic==0.4.27
Pillow==10.2.0

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# AI and Machine Learning dependencies
torch==2.2.0
transformers==4.38.2
accelerate==0.27.2
peft==0.9.0
bitsandbytes==0.42.0

# Speech recognition (Whisper)
openai-whisper==20231117
faster-whisper==1.0.1

# LLM dependencies
llama-cookbook==0.0.5
huggingface-hub==0.20.3
safetensors==0.4.2

# Configuration and environment
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities and helpers
python-dateutil==2.8.2
pytz==2024.1
uuid==1.30
requests==2.31.0

# Logging and monitoring
python-json-logger==2.0.7
rich==13.7.0

# Development and testing
pytest==8.0.2
psutil==5.9.8

# Data processing
numpy==1.26.4
pandas==2.2.0

# Optional: Database drivers (uncomment as needed)
# psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0          # MySQL 